Тренд на cogs/tools.py'ему раздробленность запущен
Многие команды были перенесены в отдельные файлы для облегчения работы.

Убраны команды: /outages, /version.

Окончательно удалены неработающие строки кода команды /qrcode.

Возвращена модерация в /help.

В /botinfo теперь используется библиотека distro.
